be/services/leaderBoardServices/requirements.txt
flask
flask_cors
redis
python-dotenv
gevent
requests
